+++ client/src/constants.ts
@@ -17,7 +17,7 @@ export const LS_VENV_NAME = "textxls";
 export const LS_WHEELS_DIR = "wheels";
 export const TEXTX_LS_CORE = "textx_ls_core";
 export const TEXTX_LS_SERVER = "textx_ls_server";
-export const VSCE_COMMAND_PATH = join(EXTENSION_PATH, "libs", "node_modules", "vsce", "out", "vsce"); // tslint:disable: max-line-length
+export const VSCE_COMMAND_PATH = join(EXTENSION_PATH, "libs", "node_modules", "vsce", "out", "vsce");
 
 // Commands
 // TextX-LS
