+++ client/src/services/index.ts
@@ -1,0 +1,6 @@
+export { IEventService, EventService } from "./eventService";
+export { IExtensionService, ExtensionService } from "./extensionService";
+export { IGeneratorService, GeneratorService } from "./generatorService";
+export { IProjectService, ProjectService } from "./projectService";
+export { ISyntaxHighlightService, SyntaxHighlightService } from "./syntaxHighlightService";
+export { IWatcherService, WatcherService } from "./watcherService";
