+++ client/tslint.json
@@ -1,50 +1,45 @@
 {
-    "extends": "tslint:recommended",
-    "rulesDirectory": [
-        "./node_modules/tslint-microsoft-contrib"
+  "extends": "tslint:recommended",
+  "rulesDirectory": [
+    "./node_modules/tslint-microsoft-contrib"
+  ],
+  "rules": {
+    "no-console": [
+      false,
+      "error",
+      "log",
+      "warn"
     ],
-    "rules": {
-        "max-line-length": {
-            "options": [
-                100
-            ]
-        },
-        "no-console": [
-            false,
-            "error",
-            "log",
-            "warn"
+    "no-unused-expression": true,
+    "typedef": [
+      false,
+      "arrow-call-signature",
+      "arrow-parameter",
+      "call-signature",
+      "member-variable-declaration",
+      "parameter",
+      "property-declaration",
+      "variable-declaration"
+    ],
+    "class-name": true,
+    "no-unexternalized-strings": [
+      false,
+      {
+        "messageIndex": 1,
+        "signatures": [
+          "localize",
+          "nls.localize"
         ],
-        "no-unused-expression": true,
-        "typedef": [
-            false,
-            "arrow-call-signature",
-            "arrow-parameter",
-            "call-signature",
-            "member-variable-declaration",
-            "parameter",
-            "property-declaration",
-            "variable-declaration"
-        ],
-        "class-name": true,
-        "no-unexternalized-strings": [
-            false,
-            {
-                "messageIndex": 1,
-                "signatures": [
-                    "localize",
-                    "nls.localize"
-                ],
-                "keyIndex": 0
-            }
-        ],
-        "semicolon": [
-            true,
-            "always"
-        ],
-        "curly": true,
-        "no-duplicate-variable": true,
-        "triple-equals": true,
-        "radix": false
-    }
+        "keyIndex": 0
+      }
+    ],
+    "semicolon": [
+      true,
+      "always"
+    ],
+    "curly": true,
+    "no-duplicate-variable": true,
+    "triple-equals": true,
+    "radix": false
+  }
 }
